requirements.txt
asn1crypto==0.24.0
backports.functools-lru-cache==1.5
bcrypt==3.1.4
boto==2.48.0
boto3==1.7.24
botocore==1.10.24
bz2file==0.98
cachetools==2.0.1
certifi==2018.1.18
cffi==1.11.5
chardet==3.0.4
cryptography==2.2.2
cycler==0.10.0
docutils==0.14
enum34==1.1.6
futures==3.2.0
gensim==3.4.0
google-api-core==1.1.0
google-auth==1.4.1
google-cloud-core==0.28.1
google-cloud-storage==1.8.0
google-cloud-translate==1.3.1
google-resumable-media==0.3.1
googleapis-common-protos==1.5.3
idna==2.6
ipaddress==1.0.19
jmespath==0.9.3
kiwisolver==1.0.1
matplotlib==2.2.2
nltk==3.3
numpy==1.14.2
pandas==0.22.0
paramiko==2.4.1
pathlib==1.0.1
protobuf==3.5.2.post1
pyasn1==0.4.2
pyasn1-modules==0.2.1
pycparser==2.18
PyNaCl==1.2.1
pyparsing==2.2.0
python-dateutil==2.7.0
pytz==2018.3
requests==2.18.4
rsa==3.4.2
s3transfer==0.1.13
scipy==1.1.0
six==1.11.0
smart-open==1.5.7
stop-words==2015.2.23.1
storage==0.0.4.3
subprocess32==3.5.1
urllib3==1.22
